platforms: debug.dockerfile: Add user environment variables to sshd
C/C++, .NET and ASP.NET templates use ssh connection to connect to gdb
and vsdbg. sshd do not use the same user environment then the ENV
entries created by the extension do not exist for debugging environment.
This commit add the "PermitUserEnvironment yes" on the sshd_config and
make a copy on the user environment variable to the /etc/environment,
so all the new ssh connections will have the same environment.

Related-to: TIE-14
Signed-off-by: Matheus Castello <matheus.castello@toradex.com>