moses: platforms: c-cpp: Fix remote machine required tools warning
From VS 2019 16.8.1 we begin to receive the warning: "A remote machine
is missing required tools. This could cause failures when building or
debugging." We can get rid of this message by adding the mentioned
packages.

Also for the case of g++ we are adding a empty executable file for the
debug container. g++ will increase container size by many megabytes and
will provide no useful features for the target, other than removing the
warning.

Related-to: TIE-282
Signed-off-by: Matheus Castello <matheus.castello@toradex.com>
Suggested-by: Valter Minute <valter.minute@toradex.com>