add reboot and shutdown requests on TargetDevice
In this commit there will be added a reboot and a shutdown function on
the TargetDevice of the csharp API to allow the frontend vs-plugin to
see and use the reboot and shutdown requests.
This function will also update the state of the device, so that it
gets updated on the device UI on Visual Studio. The shutdown will show
a "Device Disconnected" message and the reboot will show the loading
bar (ConnectonState.Unknown). The reboot boolean is added to the
RefreshAsync interface because it is necessary to differentiate when
the device is disconnected while rebooting and when the device is
just disconnected (requiring a user refresh to reconnect).

Signed-off-by Andre Riesco <andre.riesco@toradex.com>